refactor(validationTest): ValidationTest 패키지 분리
test
-validation
--MoneyValidationTest
--WinningNumbersValidationTest
--BonusNumberValidationTest
